requirements.txt
pandas
dash
plotly
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
